+++ src/main/java/doConnect/cogentcapstone/controller/ChatController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import doConnect.cogentcapstone.entity.Chat;
 import doConnect.cogentcapstone.repository.ChatRepository;
 import doConnect.cogentcapstone.service.ChatService;
+import java.util.ArrayList;
 
 @RestController
 @CrossOrigin("*")
@@ -26,26 +27,48 @@ public class ChatController {
 	
 	// TODO: move this functionality to AnswerService class
 	@Autowired
-	private ChatRepository chatRepository; 
-	@Autowired
-	private ChatService chatService;
+	private ChatService cts;
 
 	@PostMapping("addMessage")
 	public Chat addMessage(@Validated @RequestBody Chat chat)
 	{
-		return chatRepository.save(chat);
+		return cts.update(chat);
 	}
 	
 	@DeleteMapping("deleteChatById/{id}")
 	public void deleteChatById(@PathVariable("id") Integer id)
 	{
-		chatRepository.deleteById(id);
+		cts.deleteByID(id);
                 System.out.println("deleted.");
 	}
 	
-	@GetMapping("getAllChatMessages")
-	public List<Chat> getAllChatMessages()
+        
+	@GetMapping("getAllMessagesBetweenUsers/{userName1}/{userName2}")
+	public List<Chat> getAllMessagesBetweenUsers(@PathVariable("userName1") String userName1, @PathVariable("userName2") String userName2)
 	{
-		return chatRepository.findAll();
+            List<Chat> from_user1 = cts.getByfromUser(userName1);
+            List<Chat> from_user2 = cts.getByfromUser(userName2);
+            List<Chat> to_user1 = cts.getBytoUser(userName1);
+            List<Chat> to_user2 = cts.getBytoUser(userName2);
+            ArrayList<Chat> all = new ArrayList<>();
+            
+            for (Chat c : from_user1) {
+                if (c.getToUser().equals(userName2)) {
+                    all.add(c);
+                }
+            }
+            
+            for (Chat c : from_user2) {
+                if (c.getToUser().equals(userName1)) {
+                    all.add(c);
+                }
+            }
+                
+            return all;
 	}
+
+        @GetMapping("getAllChatMessages")
+        public List<Chat> getAllChatMessages() {
+            return cts.getAllChatMessages();
+        }
 }
