+++ src/main/java/doConnect/cogentcapstone/repository/AnswerRepository.java
@@ -3,7 +3,6 @@ package doConnect.cogentcapstone.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import doConnect.cogentcapstone.entity.Answer;
-import doConnect.cogentcapstone.entity.Question;
 import java.util.List;
 
 @Repository
